Rename ParserWorker to RevealParserWorker (#722)
More importantly renames its corresponding .worker.js file from `parser.worker.js` to `reveal.parser.worker.js` and the web worker from `parser`to `reveal.parser`. This avoids any potential conflicts with other packages.
