+++ viewer/src/datamodels/cad/sector/CadSectorParser.ts
@@ -3,8 +3,8 @@
  */
 
 import { WorkerPool } from '@/utilities/workers/WorkerPool';
-import { ParseSectorResult, ParseCtmResult } from '@/utilities/workers/types/parser.types';
-import { ParserWorker } from '@/utilities/workers/parser.worker';
+import { ParseSectorResult, ParseCtmResult } from '@/utilities/workers/types/reveal.parser.types';
+import { RevealParserWorker } from '@/utilities/workers/reveal.parser.worker';
 import { SectorQuads } from '../rendering/types';
 
 export class CadSectorParser {
@@ -26,16 +26,18 @@ export class CadSectorParser {
   }
 
   private async parseSimple(quadsArrayBuffer: Uint8Array): Promise<SectorQuads> {
-    return this.workerPool.postWorkToAvailable<SectorQuads>(async (worker: ParserWorker) =>
+    return this.workerPool.postWorkToAvailable<SectorQuads>(async (worker: RevealParserWorker) =>
       worker.parseQuads(quadsArrayBuffer)
     );
   }
 
   private async parseDetailed(sectorArrayBuffer: Uint8Array): Promise<ParseSectorResult> {
-    return this.workerPool.postWorkToAvailable(async (worker: ParserWorker) => worker.parseSector(sectorArrayBuffer));
+    return this.workerPool.postWorkToAvailable(async (worker: RevealParserWorker) =>
+      worker.parseSector(sectorArrayBuffer)
+    );
   }
 
   private async parseCtm(ctmArrayBuffer: Uint8Array): Promise<ParseCtmResult> {
-    return this.workerPool.postWorkToAvailable(async (worker: ParserWorker) => worker.parseCtm(ctmArrayBuffer));
+    return this.workerPool.postWorkToAvailable(async (worker: RevealParserWorker) => worker.parseCtm(ctmArrayBuffer));
   }
 }
