+++ viewer/src/utilities/workers/types/reveal.parser.types.ts
